web-component for custom-data-list
