+++ src-js/views-fontinfo/src/panel-sources.js
@@ -24,10 +24,12 @@ import {
   mapAxesFromUserSpaceToSourceSpace,
 } from "@fontra/core/var-model.js";
 import "@fontra/web-components/add-remove-buttons.js";
+import "@fontra/web-components/custom-data-list.js";
+import { CustomDataList } from "@fontra/web-components/custom-data-list.js";
 import "@fontra/web-components/designspace-location.js";
 import { dialogSetup, message } from "@fontra/web-components/modal-dialog.js";
 import { UIList } from "@fontra/web-components/ui-list.js";
-import { arraysEqual, updateRemoveButton } from "./panel-axes.js";
+import { updateRemoveButton } from "./panel-axes.js";
 import { BaseInfoPanel } from "./panel-base.js";
 
 let selectedSourceIdentifier = undefined;
@@ -753,11 +755,15 @@ class SourceBox extends HTMLElement {
         ]),
         buildFontGuidelineList(this.controllers.guidelines)
       );
+      const customDataList = new CustomDataList({
+        controller: this.controllers.customData,
+        fontObject: this.source,
+      });
       this.append(
         html.div({ class: "fontra-ui-font-info-sources-panel-header" }, [
           getLabelFromKey("customData"),
         ]),
-        buildFontCustomDataList(this.controllers.customData, this.source)
+        customDataList
       );
     }
   }
@@ -933,138 +939,6 @@ function buildFontGuidelineList(controller) {
   updateRemoveButton(labelList, addRemoveButton);
 
   return html.div({ style: "display: grid; grid-gap: 0.3em; padding-bottom: 2em;" }, [
-    labelList,
-    addRemoveButton,
-  ]);
-}
-
-export function buildFontCustomDataList(
-  controller,
-  fontObject = undefined,
-  supportedAttributes = Object.keys(customDataNameMapping)
-) {
-  // fontObject can either be FontInfo or FontSource.
-  const customDataNames = Object.keys(customDataNameMapping);
-  const model = controller.model;
-
-  const makeItem = ([key, value]) => {
-    const item = new ObservableController({ key: key, value: value });
-    item.addListener((event) => {
-      const sortedItems = [...labelList.items];
-      sortedItems.sort(
-        (a, b) =>
-          (customDataNames.indexOf(a.key) != -1
-            ? customDataNames.indexOf(a.key)
-            : customDataNames.length) -
-          (customDataNames.indexOf(b.key) != -1
-            ? customDataNames.indexOf(b.key)
-            : customDataNames.length)
-      );
-
-      if (!arraysEqual(labelList.items, sortedItems)) {
-        labelList.setItems(sortedItems);
-      }
-
-      const newCustomData = sortedItems.map((customData) => {
-        return { ...customData };
-      });
-      model.customData = newCustomData;
-    });
-    return item.model;
-  };
-
-  const sortedItems = Object.entries(model);
-  sortedItems.sort(
-    (a, b) =>
-      (customDataNames.indexOf(a[0]) != -1
-        ? customDataNames.indexOf(a[0])
-        : customDataNames.length) -
-      (customDataNames.indexOf(b[0]) != -1
-        ? customDataNames.indexOf(b[0])
-        : customDataNames.length)
-  );
-  const items = sortedItems?.map(makeItem) || [];
-
-  const labelList = new UIList();
-  labelList.classList.add("fontra-ui-font-info-sources-panel-list-element");
-  labelList.style = `min-width: 12em;`;
-  labelList.columnDescriptions = [
-    {
-      key: "key",
-      title: "Key", // TODO: translation
-      width: "14em",
-      editable: true,
-      continuous: false,
-    },
-    {
-      key: "value",
-      title: "Value", // TODO: translation
-      width: "10em",
-      editable: true,
-      continuous: false,
-    },
-  ];
-  labelList.showHeader = true;
-  labelList.minHeight = "5em";
-  labelList.setItems(items);
-
-  const deleteSelectedItem = () => {
-    const index = labelList.getSelectedItemIndex();
-    if (index === undefined) {
-      return;
-    }
-    const items = [...labelList.items];
-    items.splice(index, 1);
-    labelList.setItems(items);
-    const newCustomData = items.map((customData) => {
-      return { ...customData };
-    });
-    model.customData = newCustomData;
-    addRemoveButton.scrollIntoView({
-      behavior: "auto",
-      block: "nearest",
-      inline: "nearest",
-    });
-    labelList.setSelectedItemIndex(items.length - 1);
-  };
-
-  labelList.addEventListener("deleteKey", deleteSelectedItem);
-  const addRemoveButton = html.createDomElement("add-remove-buttons", {
-    addButtonCallback: () => {
-      // TODO: Maybe open a dialog with a list of possible keys?
-      const currentKeys = labelList.items.map((customData) => {
-        return customData.key;
-      });
-      let nextKey = "attributeName";
-      for (const key of supportedAttributes) {
-        if (!currentKeys.includes(key)) {
-          nextKey = key;
-          break;
-        }
-      }
-      const valueDefault = customDataNameMapping[nextKey]
-        ? customDataNameMapping[nextKey].default(fontObject)
-        : "";
-      const newItem = makeItem([nextKey, valueDefault]);
-      const newItems = [...labelList.items, newItem];
-      model.customData = newItems.map((label) => {
-        return { ...label };
-      });
-      labelList.setItems(newItems);
-      labelList.editCell(newItems.length - 1, "key");
-      addRemoveButton.scrollIntoView({
-        behavior: "auto",
-        block: "nearest",
-        inline: "nearest",
-      });
-    },
-    removeButtonCallback: deleteSelectedItem,
-    disableRemoveButton: true,
-  });
-
-  updateRemoveButton(labelList, addRemoveButton);
-
-  return html.div({ style: "display: grid; grid-gap: 0.3em;" }, [
     labelList,
     addRemoveButton,
   ]);
